Rob, check out this. File Tree UI now works, at least on my computer. Let me know if that works on yours., too.
Progress:
a.Can create both Seq or Asset as reference 
Just click on Asset name or Type's name, click on seq or shot's names won't work, you can create a reference of it.

b.Popup menu add.
Right click on the reference item, you get
1. select contents
2. remove it
3. import

In the furture:
a. more function for reference can be made.
b. function that open/close/save/delete.... seq.
c. son on and so on .....